requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.2
cssutils==1.0.2
dj-database-url==0.5.0
Django==3.1.1
django-bootstrap-modal-forms==2.0.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-heroku==0.3.1
django-inlinecss==0.3.0
django-tables2==2.3.1
future==0.18.2
gunicorn==20.0.4
mysqlclient==2.0.1
psycopg2==2.8.6
pynliner==0.8.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.2.0
